backend/pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
target-version = [
    "py36",
    "py37",
    "py38",
    "py39",
    "py310",
]
exclude = '''
/(
    \.eggs
  | \.git
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
default_section = "THIRDPARTY"
known_first_party = "core,django_search_example,gutensearch,scripts,tests"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-rA"
]
testpaths = [
    "tests",
]

[tool.poetry]
name = "django-search-example"
version = "0.1.0"
description = ""
authors = ["Thomas Aglassinger <roskakori@users.sourceforge.net>"]

[tool.poetry.dependencies]
python = ">=3.8, <4"
Django = "^4.0.4"
django-extensions = "^3.1.5"
django-debug-toolbar = "^3.4.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
docker-compose = "^1.29.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
pre-commit = "^2.19.0"

[pycodestyle]
max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
